saving checkpoint, working collection whitelist setup in test
