+++ contracts/sg-eth-airdrop/src/tests/constants.rs
@@ -1,0 +1,7 @@
+pub const OWNER: &str = "admin0001";
+pub const AIRDROP_CONTRACT: &str = "contract0";
+pub const STARGAZE_WALLET_01: &str = "0xstargaze_wallet_01";
+pub const STARGAZE_WALLET_02: &str = "0xstargaze_wallet_02";
+pub const CONTRACT_CONFIG_PLAINTEXT: &str =
+    "My Stargaze address is {wallet} and I want a Winter Pal.";
+pub const NATIVE_DENOM: &str = "ustars";
