+++ contracts/sg-eth-airdrop/src/tests/lib.rs
@@ -1,0 +1,8 @@
+mod constants;
+mod shared;
+#[cfg(test)]
+pub mod test_claim;
+#[cfg(test)]
+pub mod test_collection_whitelist;
+#[cfg(test)]
+pub mod test_eth_whitelist;
