+++ contracts/sg-eth-airdrop/src/tests/test_collection_whitelist.rs
@@ -1,33 +1,16 @@
-use async_std::task;
-use cosmwasm_std::{Addr, Attribute, Timestamp, coin, coins};
-use cw_multi_test::error::Error;
-use cw_multi_test::{AppResponse, Contract, ContractWrapper, Executor, SudoMsg, BankSudo};
-use ethers_core::k256::ecdsa::SigningKey;
-use ethers_core::types::H160;
+use cosmwasm_std::{coin, coins, Addr, Timestamp};
+use cw_multi_test::{BankSudo, Contract, ContractWrapper, Executor, SudoMsg};
 
 use sg2::msg::Sg2ExecuteMsg;
 use sg_multi_test::StargazeApp;
 use sg_std::{self, StargazeMsgWrapper, GENESIS_MINT_START_TIME, NATIVE_DENOM};
 
-// use sg_std::StargazeMsgWrapper;
-use std::str;
-
 use crate::contract::reply;
-use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
-
-use ethers_core::rand::thread_rng;
-use ethers_signers::{LocalWallet, Signer, Wallet, WalletError};
-use eyre::Result;
 use sg2::tests::mock_collection_params;
 use vending_factory::msg::{VendingMinterCreateMsg, VendingMinterInitMsgExtension};
 use vending_factory::state::{ParamsExtension, VendingMinterParams};
 
 extern crate whitelist_generic;
-
-const OWNER: &str = "admin0001";
-const AIRDROP_CONTRACT: &str = "contract0";
-const STARGAZE_WALLET_01: &str = "0xstargaze_wallet_01";
-const CONTRACT_CONFIG_PLAINTEXT: &str = "My Stargaze address is {wallet} and I want a Winter Pal.";
 
 const CREATION_FEE: u128 = 5_000_000_000;
 const INITIAL_BALANCE: u128 = 2_000_000_000;
@@ -50,15 +33,15 @@ fn custom_mock_app() -> StargazeApp {
     StargazeApp::default()
 }
 
-// pub fn contract() -> Box<dyn Contract<sg_std::StargazeMsgWrapper>> {
-//     let contract = ContractWrapper::new(
-//         crate::contract::execute,
-//         crate::contract::instantiate,
-//         crate::contract::query,
-//     )
-//     .with_reply(reply);
-//     Box::new(contract)
-// }
+pub fn contract() -> Box<dyn Contract<sg_std::StargazeMsgWrapper>> {
+    let contract = ContractWrapper::new(
+        crate::contract::execute,
+        crate::contract::instantiate,
+        crate::contract::query,
+    )
+    .with_reply(reply);
+    Box::new(contract)
+}
 
 pub fn whitelist_generic_contract() -> Box<dyn Contract<StargazeMsgWrapper>> {
     let contract = ContractWrapper::new(
@@ -78,7 +61,6 @@ pub fn contract_minter() -> Box<dyn Contract<StargazeMsgWrapper>> {
     .with_reply(vending_minter::contract::reply);
     Box::new(contract)
 }
-
 
 pub fn contract_whitelist() -> Box<dyn Contract<StargazeMsgWrapper>> {
     let contract = ContractWrapper::new(
@@ -107,8 +89,6 @@ pub fn contract_sg721() -> Box<dyn Contract<StargazeMsgWrapper>> {
     Box::new(contract)
 }
 
-
-
 fn setup_whitelist_contract(router: &mut StargazeApp, creator: &Addr) -> Addr {
     let whitelist_code_id = router.store_code(contract_whitelist());
 
@@ -131,42 +111,6 @@ fn setup_whitelist_contract(router: &mut StargazeApp, creator: &Addr) -> Addr {
         )
         .unwrap()
 }
-
-
-// fn get_instantiate_contract(addresses: Vec<String>, funds_amount: u128) -> StargazeApp {
-//     let mut app = custom_mock_app();
-//     app.sudo(SudoMsg::Bank({
-//         BankSudo::Mint {
-//             to_address: OWNER.to_string(),
-//             amount: coins(funds_amount, NATIVE_DENOM),
-//         }
-//     }))
-//     .map_err(|err| println!("{:?}", err))
-//     .ok();
-
-//     let sg_eth_id = app.store_code(contract());
-//     let whitelist_code_id = app.store_code(whitelist_generic_contract());
-//     assert_eq!(sg_eth_id, 1);
-//     let msg: InstantiateMsg = InstantiateMsg {
-//         admin: Addr::unchecked(OWNER),
-//         claim_msg_plaintext: Addr::unchecked(CONTRACT_CONFIG_PLAINTEXT).into_string(),
-//         airdrop_amount: 3000,
-//         minter_page: "http://levana_page/airdrop".to_string(),
-//         addresses,
-//         minter_code_id: whitelist_code_id,
-//     };
-//     app.instantiate_contract(
-//         sg_eth_id,
-//         Addr::unchecked(OWNER),
-//         &msg,
-//         &coins(funds_amount, NATIVE_DENOM),
-//         "sg-eg-airdrop",
-//         Some(Addr::unchecked(OWNER).to_string()),
-//     )
-//     .unwrap();
-//     app
-// }
-
 
 pub fn mock_init_extension(splits_addr: Option<String>) -> VendingMinterInitMsgExtension {
     vending_factory::msg::VendingMinterInitMsgExtension {
@@ -204,7 +148,6 @@ pub fn mock_params() -> VendingMinterParams {
     }
 }
 
-
 // Upload contract code and instantiate minter contract
 fn setup_minter_contract(
     router: &mut StargazeApp,
@@ -232,7 +175,7 @@ fn setup_minter_contract(
             None,
         )
         .unwrap();
-    
+
     let sg721_code_id = router.store_code(contract_sg721());
     println!("sg721_code_id: {}", sg721_code_id);
 
@@ -252,7 +195,10 @@ fn setup_minter_contract(
 
     let config: vending_minter::msg::ConfigResponse = router
         .wrap()
-        .query_wasm_smart(minter_addr.clone(), &vending_minter::msg::QueryMsg::Config {})
+        .query_wasm_smart(
+            minter_addr.clone(),
+            &vending_minter::msg::QueryMsg::Config {},
+        )
         .unwrap();
 
     (minter_addr, config)
@@ -307,62 +253,17 @@ fn setup_block_time(router: &mut StargazeApp, nanos: u64, height: Option<u64>) {
     router.set_block(block);
 }
 
-async fn get_signature(
-    wallet: Wallet<SigningKey>,
-    plaintext_msg: &str,
-) -> Result<ethers_core::types::Signature, WalletError> {
-    wallet.sign_message(plaintext_msg).await
-}
-
-fn get_wallet_and_sig(
-    claim_plaintext: String,
-) -> (
-    Wallet<ethers_core::k256::ecdsa::SigningKey>,
-    std::string::String,
-    H160,
-    std::string::String,
-) {
-    let wallet = LocalWallet::new(&mut thread_rng());
-    let eth_sig_str = task::block_on(get_signature(wallet.clone(), &claim_plaintext))
-        .unwrap()
-        .to_string();
-    let eth_address = wallet.address();
-    let eth_addr_str = format!("{:?}", eth_address);
-    (wallet, eth_sig_str, eth_address, eth_addr_str)
-}
-
-fn execute_contract_with_msg(
-    msg: ExecuteMsg,
-    app: &mut StargazeApp,
-    user: Addr,
-) -> Result<AppResponse, Error> {
-    let sg_eth_addr = Addr::unchecked(AIRDROP_CONTRACT);
-
-    let result = app.execute_contract(user, sg_eth_addr, &msg, &[]).unwrap();
-    Ok(result)
-}
-
-fn get_msg_plaintext(wallet_address: String) -> String {
-    str::replace(CONTRACT_CONFIG_PLAINTEXT, "{wallet}", &wallet_address)
-}
-
-// #[test]
-// fn test_instantiate() {
-//     get_instantiate_contract(vec![], 10000);
-// }
-
-#[test]
-fn whitelist_access_len_add_remove_expiration() {
-    let mut router = custom_mock_app();
-    let (creator, buyer) = setup_accounts(&mut router);
-    let num_tokens = 1;
-    let (minter_addr, config) = setup_minter_contract(&mut router, &creator, num_tokens, None);
-    let sg721_addr = config.sg721_address;
-    let whitelist_addr = setup_whitelist_contract(&mut router, &creator);
+fn configure_collection_whitelist(
+    router: &mut StargazeApp,
+    creator: Addr,
+    buyer: Addr,
+    minter_addr: Addr,
+) -> Addr {
+    let whitelist_addr = setup_whitelist_contract(router, &creator);
     const AFTER_GENESIS_TIME: Timestamp = Timestamp::from_nanos(GENESIS_MINT_START_TIME + 100);
 
     // Set to just before genesis mint start time
-    setup_block_time(&mut router, GENESIS_MINT_START_TIME - 10, None);
+    setup_block_time(router, GENESIS_MINT_START_TIME - 10, None);
 
     // Update whitelist_expiration fails if not admin
     let wl_msg = sg_whitelist::msg::ExecuteMsg::UpdateEndTime(AFTER_GENESIS_TIME);
@@ -389,6 +290,52 @@ fn whitelist_access_len_add_remove_expiration() {
         &set_whitelist_msg,
         &[],
     );
-    // println!("response is {:?}", res.unwrap());
+    assert!(res.is_ok());
+    whitelist_addr
+}
+
+#[test]
+fn whitelist_access_len_add_remove_expiration() {
+    let mut router = custom_mock_app();
+    let (creator, buyer) = setup_accounts(&mut router);
+    let num_tokens = 1;
+    let (minter_addr, config) = setup_minter_contract(&mut router, &creator, num_tokens, None);
+    let sg721_addr = config.sg721_address;
+
+    let whitelist_addr = configure_collection_whitelist(
+        &mut router,
+        creator.clone(),
+        buyer.clone(),
+        minter_addr.clone(),
+    );
+
+    // Mint fails, buyer is not on whitelist
+    let mint_msg = vending_minter::msg::ExecuteMsg::Mint {};
+    let res = router.execute_contract(
+        buyer.clone(),
+        minter_addr.clone(),
+        &mint_msg,
+        &coins(MINT_PRICE, NATIVE_DENOM),
+    );
+    assert!(res.is_err());
+
+    // Add buyer to whitelist
+    let inner_msg = sg_whitelist::msg::AddMembersMsg {
+        to_add: vec![buyer.to_string()],
+    };
+    let wasm_msg = sg_whitelist::msg::ExecuteMsg::AddMembers(inner_msg);
+    let res = router.execute_contract(creator.clone(), whitelist_addr.clone(), &wasm_msg, &[]);
+    assert!(res.is_ok());
+
+    setup_block_time(&mut router, GENESIS_MINT_START_TIME, None);
+
+    // Mint succeeds with whitelist price
+    let mint_msg = vending_minter::msg::ExecuteMsg::Mint {};
+    let res = router.execute_contract(
+        buyer.clone(),
+        minter_addr.clone(),
+        &mint_msg,
+        &coins(WHITELIST_AMOUNT, NATIVE_DENOM),
+    );
     assert!(res.is_ok());
 }
